+++ threedi_modelchecker/simulation_templates/structure_controls/extractor.py
@@ -1,3 +1,4 @@
+from datetime import timedelta
 from sqlalchemy.orm import Query
 from sqlalchemy.orm.session import Session
 from threedi_api_client.openapi.models.measure_location import MeasureLocation
@@ -9,9 +10,6 @@ from threedi_api_client.openapi.models.memory_structure_control import (
 )
 from threedi_api_client.openapi.models.table_structure_control import (
     TableStructureControl,
-)
-from threedi_api_client.openapi.models.timed_structure_control import (
-    TimedStructureControl,
 )
 from threedi_modelchecker.simulation_templates.exceptions import (
     SchematisationError,
@@ -26,10 +24,12 @@ from threedi_modelchecker.threedi_model.models import ControlMeasureGroup
 from threedi_modelchecker.threedi_model.models import ControlMeasureMap
 from threedi_modelchecker.threedi_model.models import ControlMemory
 from threedi_modelchecker.threedi_model.models import ControlTable
-from threedi_modelchecker.threedi_model.models import ControlTimed
 from typing import Dict
 from typing import List
 from typing import Union
+
+
+INFINITE_SIM_DURATION = int((timedelta(days=365) * 100).total_seconds())
 
 
 def control_measure_map_to_measure_location(
@@ -74,6 +74,33 @@ TYPE_MAPPING = {"set_capacity": "set_pump_capacity"}
 CAPACITY_FACTOR: float = 0.001
 
 
+def parse_action_value(value) -> List[float]:
+    # First clean the input:
+    # - remove whitespace at the start & end (strip)
+    # - replace 'internal' whitespace with ;
+    # - replace commas with ;
+    value = (";".join(value.strip().split())).replace(",", ";")
+    # This could have yielded double semicolons (in case of <whitespace>;)
+    while ";;" in value:
+        value.replace(";;", ";")
+    return [float(y) for y in value.split(";")]
+
+
+def parse_action_table(table) -> List[List[float]]:
+    """Parse an action table (for table controls)
+
+    For table controls, the action table consists of multiple records, with
+    2 or 3 values each, like:
+
+    - "1.2;2.3#1.3;2.1#1.5;5.6"
+    - "1.2;2.3;3.4#1.3;2.1;5.6#1.5;5.6;6.7"
+
+    Sometimes, the values are separated by a space and not a semicolon (;). This
+    function also accepts those inputs.
+    """
+    return [parse_action_value(x) for x in table.split("#")]
+
+
 def to_table_control(
     control: Control,
     table_control: ControlTable,
@@ -83,14 +110,10 @@ def to_table_control(
     action_type: str = TYPE_MAPPING.get(
         table_control.action_type, table_control.action_type
     )
-    # Note: Yes, table control really uses # and ;
     try:
-        values = [
-            [float(y) for y in x.split(";")]
-            for x in table_control.action_table.split("#")
-        ]
+        values = parse_action_table(table_control.action_table)
         if table_control.action_type == "set_capacity":
-            values[1] = [x * CAPACITY_FACTOR for x in values[1]]
+            values = [[x[0], x[1] * CAPACITY_FACTOR] for x in values]
     except (ValueError, TypeError):
         raise SchematisationError(
             f"Table control action_table incorrect format for v2_control_table.id = {table_control.id}"
@@ -102,15 +125,13 @@ def to_table_control(
         control_start = 0
 
     try:
-        int(control.end)
+        control_end = int(control.end)
     except (ValueError, TypeError):
-        raise SchematisationError(
-            f"Timed control control.end not set for v2_control_table.id = {table_control.id}"
-        )
+        control_end = INFINITE_SIM_DURATION
 
     return TableStructureControl(
         offset=control_start,
-        duration=int(control.end) - control_start,
+        duration=control_end - control_start,
         measure_specification=measure_specification,
         structure_id=table_control.target_id,
         structure_type=table_control.target_type,
@@ -130,7 +151,9 @@ def to_memory_control(
     )
 
     try:
-        value = [float(x) for x in memory_control.action_value.split(" ")]
+        value = parse_action_value(memory_control.action_value)
+        if memory_control.action_type == "set_capacity":
+            value = [value[0] * CAPACITY_FACTOR]
     except (ValueError, TypeError):
         raise SchematisationError(
             f"Memory control action_value incorrect format for v2_control_memory.id = {memory_control.id}"
@@ -142,15 +165,13 @@ def to_memory_control(
         control_start = 0
 
     try:
-        int(control.end)
+        control_end = int(control.end)
     except (ValueError, TypeError):
-        raise SchematisationError(
-            f"Timed control control.end not set for v2_control_memory.id = {memory_control.id}"
-        )
+        control_end = INFINITE_SIM_DURATION
 
     return MemoryStructureControl(
         offset=control_start,
-        duration=int(control.end) - control_start,
+        duration=control_end - control_start,
         measure_specification=measure_specification,
         structure_id=memory_control.target_id,
         structure_type=memory_control.target_type,
@@ -163,53 +184,6 @@ def to_memory_control(
     )
 
 
-def to_timed_control(
-    control: Control, timed_control: ControlTimed
-) -> TimedStructureControl:
-
-    action_type: str = TYPE_MAPPING.get(
-        timed_control.action_type, timed_control.action_type
-    )
-
-    try:
-        values = [
-            [float(y) for y in x.split(";")]
-            for x in timed_control.action_table.split("#")
-        ]
-
-        if timed_control.action_type == "set_capacity":
-            values[1] = [x * CAPACITY_FACTOR for x in values[1]]
-
-    except (ValueError, TypeError):
-        raise SchematisationError(
-            f"Timed control action_table incorrect format for v2_control_timed.id = {timed_control.id}"
-        )
-
-    try:
-        control_start = int(control.start)
-    except (ValueError, TypeError):
-        control_start = 0
-
-    try:
-        int(control.end)
-    except (ValueError, TypeError):
-        raise SchematisationError(
-            f"Timed control control.end not set for v2_control_timed.id = {timed_control.id}"
-        )
-
-    # Pick first two values
-    value = values[0]
-
-    return TimedStructureControl(
-        offset=control_start,
-        duration=int(control.end) - control_start,
-        value=value,
-        type=action_type,
-        structure_id=timed_control.target_id,
-        structure_type=timed_control.target_type,
-    )
-
-
 class StructureControlExtractor(object):
     def __init__(self, session: Session, control_group_id: int):
         self.session = session
@@ -218,7 +192,7 @@ class StructureControlExtractor(object):
 
     def __initialize_controls(self):
         if self._controls is None:
-            self._controls = {"timed": [], "table": [], "memory": []}
+            self._controls = {"table": [], "memory": []}
             table_lookup = dict(
                 [
                     (x.id, x)
@@ -229,12 +203,6 @@ class StructureControlExtractor(object):
                 [
                     (x.id, x)
                     for x in Query(ControlMemory).with_session(self.session).all()
-                ]
-            )
-            timed_lookup = dict(
-                [
-                    (x.id, x)
-                    for x in Query(ControlTimed).with_session(self.session).all()
                 ]
             )
             maps_lookup = {}
@@ -275,20 +243,6 @@ class StructureControlExtractor(object):
                     continue
                 self._controls[control.control_type].append(api_control)
 
-            for control in (
-                Query(Control)
-                .filter(
-                    Control.control_type == "timed",
-                    Control.control_group_id == self._control_group_id,
-                )
-                .with_session(self.session)
-                .all()
-            ):
-
-                timed: ControlTimed = timed_lookup[control.control_id]
-                api_control = to_timed_control(control, timed)
-                self._controls[control.control_type].append(api_control)
-
     def all_controls(self) -> StructureControls:
         self.__initialize_controls()
         return StructureControls(**self._controls)
@@ -297,11 +251,6 @@ class StructureControlExtractor(object):
     def memory_controls(self) -> List[MemoryStructureControl]:
         self.__initialize_controls()
         return self._controls["memory"]
-
-    @property
-    def timed_controls(self) -> List[TimedStructureControl]:
-        self.__initialize_controls()
-        return self._controls["timed"]
 
     @property
     def table_controls(self) -> List[TableStructureControl]: