+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changelog of threedi-modelchecker
 0.20 (unreleased)
 -----------------
 
+- Bugfix: Correct parsing of discharge_coefficients in table control.
+
+- Removed timed control parsing, there are no models using it.
+
 - Set maximum_time_step to sim_time_step if maximum_time_step is NULL or
   less than sim_time_step. Reverts change in 0.19.